+++ src/containers/recipe_form_container.js
@@ -1,5 +1,5 @@
-import './add_recipe_container.css';
-import { RecipeForm } from '../components/add_recipe_form';
+import './recipe_form_container.css';
+import { RecipeForm } from '../components/recipe_form';
 
 export const AddRecipeContainer = ({ setDivVisible }) => {
   return (
